Move unit test to IntermediateTypeInspectionResultSpec
