+++ modules/test-it/src/test/scala/org/knora/webapi/messages/util/search/gravsearch/types/GravsearchTypeInspectorSpec.scala
@@ -1310,71 +1310,6 @@ class GravsearchTypeInspectorSpec extends E2ESpec {
   }
 
   "The inferring type inspector" should {
-    "remove a type from IntermediateTypeInspectionResult" in {
-      val multipleDetectedTypes: IntermediateTypeInspectionResult = IntermediateTypeInspectionResult(
-        entities = Map(
-          TypeableVariable(variableName = "mainRes") -> Set(
-            NonPropertyTypeInfo(
-              typeIri = "http://api.knora.org/ontology/knora-api/simple/v2#Resource".toSmartIri,
-              isResourceType = true,
-            ),
-            NonPropertyTypeInfo(
-              typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#letter".toSmartIri,
-              isResourceType = true,
-            ),
-            NonPropertyTypeInfo(
-              typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#basicLetter".toSmartIri,
-              isResourceType = true,
-            ),
-          ),
-        ),
-        entitiesInferredFromPropertyIris = Map(
-          TypeableVariable(variableName = "mainRes") -> Set(
-            NonPropertyTypeInfo(
-              typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#letter".toSmartIri,
-              isResourceType = true,
-            ),
-            NonPropertyTypeInfo(
-              typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#basicLetter".toSmartIri,
-              isResourceType = true,
-            ),
-          ),
-        ),
-      )
-
-      // remove type basicLetter
-      val intermediateTypesWithoutResource: IntermediateTypeInspectionResult = multipleDetectedTypes.removeType(
-        TypeableVariable(variableName = "mainRes"),
-        NonPropertyTypeInfo(
-          typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#basicLetter".toSmartIri,
-          isResourceType = true,
-        ),
-      )
-
-      // Is it removed from entities?
-      intermediateTypesWithoutResource.entities should contain(
-        TypeableVariable(variableName = "mainRes") -> Set(
-          NonPropertyTypeInfo(
-            typeIri = "http://api.knora.org/ontology/knora-api/simple/v2#Resource".toSmartIri,
-            isResourceType = true,
-          ),
-          NonPropertyTypeInfo(
-            typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#letter".toSmartIri,
-            isResourceType = true,
-          ),
-        ),
-      )
-
-      // Is it removed from entitiesInferredFromProperties?
-      intermediateTypesWithoutResource.entitiesInferredFromPropertyIris should contain(
-        TypeableVariable(variableName = "mainRes") -> Set(
-          NonPropertyTypeInfo(
-            typeIri = "http://0.0.0.0:3333/ontology/0801/beol/simple/v2#letter".toSmartIri,
-            isResourceType = true,
-          ),
-        ),
-      )
-    }
 
     "refine the inspected types for each typeableEntity" in {
       val typeInspectionRunner = UnsafeZioRun.service[InferringGravsearchTypeInspector]
